+++ Classes/Domain/Repository/TeaRepository.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace TTN\Tea\Domain\Repository\Product;
+namespace TTN\Tea\Domain\Repository;
 
-use TTN\Tea\Domain\Model\Product\Tea;
+use TTN\Tea\Domain\Model\Tea;
 use TTN\Tea\Domain\Repository\Traits\StoragePageAgnosticTrait;
 use TYPO3\CMS\Extbase\Persistence\QueryInterface;
 use TYPO3\CMS\Extbase\Persistence\QueryResultInterface;