+++ Configuration/Extbase/Persistence/Classes.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 return [
-    \TTN\Tea\Domain\Model\Product\Tea::class => [
+    \TTN\Tea\Domain\Model\Tea::class => [
         'properties' => [
             'ownerUid' => ['fieldName' => 'owner'],
         ],